Add Click on anywhere to begin game from beginGameActivity and Fix the bug which was not allowing user to exit app (#5)
* Fix Changing of what to be drawn.
Sometimes it was changing what to be drawn as reported by someone in Play Store. This is most probable reason. Do we really need to setText after clear ?

* Revert Fix Changing of what to be drawn.
Sometimes it was changing what to be drawn as reported by someone in Play Store. This is most probable reason. Do we really need to setText after clear ?

* Add click on anywhere to begin game and
Final fix for app doesn't allow user to close.

* Add confirm Exit.

* Finish the current activity before going to next one.
GameOverActivity and IntroActivity were not finished.